requirements.txt
Flask
flask-cors
scikit-learn
pandas
numpy
nltk
wordcloud
matplotlib
seaborn
gensim
transformers
torch
xgboost
datasets
Pillow
uvicorn
fastapi
python-multipart
